+++ sway-lsp/src/utils/debug.rs
@@ -1,6 +1,12 @@
-use crate::core::token::Token;
+use sway_types::Ident;
 use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity};
-use crate::core::traverse_typed_tree::{get_range_from_span, TokenMap};
+use crate::core::{
+    token::Token,
+    typed_token_type::{TokenType, TokenMap},
+    traverse_typed_tree::get_type_id
+};
+use crate::utils::common::get_range_from_span;
+
 // Flags for debugging various parts of the server
 #[derive(Debug, Default)]
 pub struct DebugFlags {
@@ -35,4 +41,48 @@ pub fn generate_warnings_for_typed_tokens(tokens: &TokenMap) -> Vec<Diagnostic> 
         .collect();
 
     warnings
+}
+
+pub fn debug_print_ident_and_token(ident: &Ident, token: &TokenType) {
+	let pos = ident.span().start_pos().line_col();
+	let line_num = pos.0 as u32;	
+    eprintln!("line num = {:?} | name: = {:?} | ast_node_type = {:?} | type_id = {:?}", 
+        line_num,
+        ident.as_str(),
+        ast_node_type(&token),
+        get_type_id(&token),
+    );
+}
+
+fn ast_node_type(token: &TokenType) -> String {
+    match &token {
+        TokenType::TypedDeclaration(dec) => {
+            dec.friendly_name().to_string()
+        }
+        TokenType::TypedExpression(exp) => {
+            exp.expression.pretty_print()
+        }
+        TokenType::TypedFunctionParameter(_) => {
+            "function parameter".to_string()
+        }
+        TokenType::TypedStructField(_) => {
+            "struct field".to_string()
+        }
+        TokenType::TypedEnumVariant(_) => {
+            "enum variant".to_string()
+        }
+        TokenType::TypedTraitFn(_) => {
+            "trait function".to_string()
+        }
+        TokenType::TypedStorageField(_) => {
+            "storage field".to_string()
+        }
+        TokenType::TypeCheckedStorageReassignDescriptor(_) => {
+            "storage reassignment descriptor".to_string()
+        }
+        TokenType::ReassignmentLhs(_) => {
+            "reassignment lhs".to_string()
+        }        
+        _ => "".to_string()
+    }
 }
